+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/markusbohl/fluent-ts-validator/branches/all?utf8=%E2%9C%93&query=feature)
 
-~~Added~~
+### Added
+- `isPostalCode()`-validation. With updating `validator.js` to `8.1.0` validation of postal codes can be performed for several different locales.
 
 ### Changed
 - `dependencies` have been updated as follows: